+++ std/encoding/csv.ts
@@ -2,7 +2,7 @@
 // https://github.com/golang/go/blob/go1.12.5/src/encoding/csv/
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
 
-import { BufReader } from "../io/bufio.ts";
+import { BufReader } from "../io/buffers.ts";
 import { TextProtoReader } from "../textproto/mod.ts";
 import { StringReader } from "../io/readers.ts";
