+++ std/io/buffers.ts
