+++ std/io/mod.ts
@@ -1,3 +1,3 @@
-export * from "./bufio.ts";
+export * from "./buffers.ts";
 export * from "./readers.ts";
 export * from "./writers.ts";
