+++ std/mime/multipart.ts
@@ -7,9 +7,9 @@ type Reader = Deno.Reader;
 type Writer = Deno.Writer;
 import { copyN, tempFile } from "../_util/io.ts";
 import { equal, findIndex, findLastIndex, hasPrefix } from "../bytes/mod.ts";
+import { BufReader, BufWriter, UnexpectedEOFError } from "../io/buffers.ts";
 import { MultiReader } from "../io/readers.ts";
 import { extname } from "../path/mod.ts";
-import { BufReader, BufWriter, UnexpectedEOFError } from "../io/bufio.ts";
 import { encoder } from "../strings/mod.ts";
 import { assertStrictEq } from "../testing/asserts.ts";
 import { TextProtoReader } from "../textproto/mod.ts";
