+++ std/ws/README.md
@@ -76,7 +76,7 @@ import {
   isWebSocketPongEvent
 } from "https://deno.land/std/ws/mod.ts";
 import { encode } from "https://deno.land/std/strings/mod.ts";
-import { BufReader } from "https://deno.land/std/io/bufio.ts";
+import { BufReader } from "https://deno.land/std/io/buffers.ts";
 import { TextProtoReader } from "https://deno.land/std/textproto/mod.ts";
 import { blue, green, red, yellow } from "https://deno.land/std/fmt/colors.ts";
