+++ std/ws/example_client.ts
@@ -1,13 +1,13 @@
+import { blue, green, red, yellow } from "../fmt/colors.ts";
+import { BufReader } from "../io/buffers.ts";
+import { encode } from "../strings/mod.ts";
+import { TextProtoReader } from "../textproto/mod.ts";
 import {
   connectWebSocket,
   isWebSocketCloseEvent,
   isWebSocketPingEvent,
   isWebSocketPongEvent
 } from "../ws/mod.ts";
-import { encode } from "../strings/mod.ts";
-import { BufReader } from "../io/bufio.ts";
-import { TextProtoReader } from "../textproto/mod.ts";
-import { blue, green, red, yellow } from "../fmt/colors.ts";
 
 const endpoint = Deno.args[1] || "ws://127.0.0.1:8080";
 /** simple websocket cli */
